server/requirements.txt
anyio==3.7.0
appdirs==1.4.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docopt==0.6.2
exceptiongroup==1.1.1
fastapi==0.96.0
ffmpeg==1.4
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
libmagic==1.0
lxml==4.9.2
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.9
pydub==0.25.1
PyJWT==2.7.0
pytest==7.3.2
python-dateutil==2.8.2
python-magic==0.4.27
python-multipart==0.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
